Moving rego code into GCV library and making audit.rego compile into the binary.
Also minor refactoring to static compile to make sure there aren't any
key collision. These map keys don't impact how the rego code is
executed, so ensuring that they keys don't collide should avoid unneeded
confusing errors.

Note this does delete the rego tests that audit used, but these are not
needed. The rego file is being tested by the CF code. This code is
actually a better test because the CF is what is defining the fields
that info is provided under.

Change-Id: I9786de2f8802e6fe891f041b2e6dafd9a683efd8