Implement ConstraintFramework.Review
* This is intended to be a functional impl of Review.  Multithreaded
support will come in a following patch.
* Fix up ConstrintFramework.Reset.  Prior to this, we weren't clearing
store.  In theory, this won't really matter since we re-create inventory
on each call to audit, but this would cause weird artifacts if people
start calling AddData then Reset then Review (which we don't really plan
to support, but we would not want to support the bugs resulting from the
weird behavior).