feat: Refactor to allow initializing validator without a file system (#155)
This splits up some of the config loading functions to allow calling them
with the contents of policy constraint and policy library files, without
reading from the file system. This makes it possible for another system using
GCV as a library to initialize a validator by directly providing config values
and not needing to read from the file system.

The semantics and usage of existing init and config loading functions are
unchanged.