Sanitize ancestry before validating an asset in ReviewAsset
SanitizeAncestry fills in the ancestors / ancestry path fields if one of them is empty; ValidateAsset requires that the ancestry path is set. This means that in practice, an ancestry path must always be set. This seems unintentional, and it means that callers like terraform-validator and cft-scorecard have to work around this behavior.